Order Controller, Service, VM, STO, Repository  Edit

+++ H2StyleStore/Models/EFModels/Order_Details.cs
@@ -21,12 +21,12 @@ namespace H2StyleStore.Models.EFModels
         [StringLength(50)]
         public string ProductName { get; set; }
 
-        public decimal UnitPrice { get; set; }
+        public int UnitPrice { get; set; }
 
-        public decimal SubTotal { get; set; }
+        public int SubTotal { get; set; }
 
         public int Quantity { get; set; }
 
-        public decimal? Discount { get; set; }
+        public int? Discount { get; set; }
     }
 }
